added an argument to linkNetwork to avoid module deletion after linking

+++ verilog/VerilogReader.cc
@@ -146,8 +146,9 @@ VerilogReader::VerilogReader(NetworkReader *network) :
   one_net_name_("one_")
 {
   network->setLinkFunc([=] (const char *top_cell_name,
-                            bool make_black_boxes) -> Instance* {
-    return linkNetwork(top_cell_name, make_black_boxes);
+                            bool make_black_boxes,
+                            bool delete_modules) -> Instance* {
+    return linkNetwork(top_cell_name, make_black_boxes, delete_modules);
   });
   constant10_max_ = stdstrPrint("%llu", std::numeric_limits<VerilogConstant10>::max());
 }
@@ -1702,7 +1703,8 @@ private:
 
 Instance *
 VerilogReader::linkNetwork(const char *top_cell_name,
-			   bool make_black_boxes)
+			   bool make_black_boxes,
+                           bool delete_modules)
 {
   if (library_) {
     Cell *top_cell = network_->findCell(library_, top_cell_name);
@@ -1730,7 +1732,9 @@ VerilogReader::linkNetwork(const char *top_cell_name,
       }
       makeModuleInstBody(module, top_instance, &bindings, make_black_boxes);
       bool errors = reportLinkErrors();
-      deleteModules();
+      if (delete_modules) {
+        deleteModules();
+      }
       if (errors) {
 	network_->deleteInstance(top_instance);
 	return nullptr;
